+++ tracker/test/tagged-events.spec.js
@@ -1,6 +1,5 @@
-const { test } = require('./support/harness')
-const { mockRequest, mockManyRequests, isMac, expectCustomEvent } = require('./support/test-utils')
-const { expect } = require('@playwright/test')
+const { mockRequest, mockManyRequests, expectCustomEvent, metaKey } = require('./support/test-utils')
+const { expect, test } = require('@playwright/test')
 
 test.describe('tagged-events extension', () => {
     test('tracks a tagged link click with custom props + url prop', async ({ page }) => {
@@ -37,14 +36,14 @@ test.describe('tagged-events extension', () => {
         expectCustomEvent(requests[0], "Form Submit", {})
     });
 
-    test('tracks click and auxclick on any tagged HTML element', async ({ page }, workerInfo) => {
+    test('tracks click and auxclick on any tagged HTML element', async ({ page }) => {
         await page.goto('/tagged-event.html')
 
         const plausibleRequestMockList = mockManyRequests(page, '/api/event', 3)
 
         await page.click('#button')
         await page.click('#span')
-        await page.click('#div', { modifiers: [isMac(workerInfo) ? 'Meta' : 'Control'] })
+        await page.click('#div', { modifiers: [metaKey()] })
 
         const requests = await plausibleRequestMockList
         expect(requests.length).toBe(3)
@@ -68,12 +67,12 @@ test.describe('tagged-events extension', () => {
         expect((await navigationRequestMock).url()).toContain(linkURL)
     });
 
-    test('tracks tagged HTML elements when their child element is clicked', async ({ page }, workerInfo) => {
+    test('tracks tagged HTML elements when their child element is clicked', async ({ page }) => {
         await page.goto('/tagged-event.html')
 
         const plausibleRequestMockList = mockManyRequests(page, '/api/event', 2)
 
-        await page.click('#h2-with-link-parent', { modifiers: [isMac(workerInfo) ? 'Meta' : 'Control'] })
+        await page.click('#h2-with-link-parent', { modifiers: [metaKey()] })
         await page.click('#link-with-div-parent')
 
         const requests = await plausibleRequestMockList